Clean up UI
Most notably:
	- Consolidate various menu scripts into one "UINavigation.cs" script\
		- This script is in the Scripts folder, where they all should have been
		- It is usally added as a component of a scene's canvas
	- Made a prefab for the pause menu, which can be dragged and dropped into the object hiearchy of a scene, and will fully function

Also experimenting with automatic level generation (as in level 3), potentially from level data files?
