--- Assets/ChangeScene.cs
@@ -1,23 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.SceneManagement;
-
-public class ChangeScene : MonoBehaviour
-{
-	public void MoveToScene(int sceneID) {
-		SceneManager.LoadScene(sceneID);
-	}
-
-    // // Start is called before the first frame update
-    // void Start()
-    // {
-        
-    // }
-
-    // // Update is called once per frame
-    // void Update()
-    // {
-        
-    // }
-}
