--- Assets/LevelsMenu.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 986d20fa8e28249e3be84562e6061f00
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
