+++ Assets/Scenes/Level Select.unity
@@ -135,6 +135,7 @@ GameObject:
   - component: {fileID: 143991625}
   - component: {fileID: 143991624}
   - component: {fileID: 143991623}
+  - component: {fileID: 143991627}
   m_Layer: 5
   m_Name: Canvas
   m_TagString: Untagged
@@ -230,6 +231,18 @@ RectTransform:
   m_AnchoredPosition: {x: 0, y: 0}
   m_SizeDelta: {x: 0, y: 0}
   m_Pivot: {x: 0, y: 0}
+--- !u!114 &143991627
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 143991622}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 14297e619d8364bdf8d2780dbd1c0762, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1 &243126836
 GameObject:
   m_ObjectHideFlags: 0
@@ -515,8 +528,8 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 851305186}
-        m_TargetAssemblyTypeName: LevelsMenu, Assembly-CSharp
+      - m_Target: {fileID: 143991627}
+        m_TargetAssemblyTypeName: UINavigation, Assembly-CSharp
         m_MethodName: ToLevel
         m_Mode: 3
         m_Arguments:
@@ -648,8 +661,8 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 851305186}
-        m_TargetAssemblyTypeName: LevelsMenu, Assembly-CSharp
+      - m_Target: {fileID: 143991627}
+        m_TargetAssemblyTypeName: UINavigation, Assembly-CSharp
         m_MethodName: ToLevel
         m_Mode: 3
         m_Arguments:
@@ -914,8 +927,8 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 851305186}
-        m_TargetAssemblyTypeName: LevelsMenu, Assembly-CSharp
+      - m_Target: {fileID: 143991627}
+        m_TargetAssemblyTypeName: UINavigation, Assembly-CSharp
         m_MethodName: ToMainMenu
         m_Mode: 1
         m_Arguments:
@@ -975,7 +988,7 @@ GameObject:
   - component: {fileID: 851305185}
   - component: {fileID: 851305184}
   - component: {fileID: 851305183}
-  - component: {fileID: 851305186}
+  - component: {fileID: 851305187}
   m_Layer: 0
   m_Name: Main Camera
   m_TagString: MainCamera
@@ -1057,7 +1070,7 @@ Transform:
   m_Children: []
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &851305186
+--- !u!114 &851305187
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -1066,9 +1079,41 @@ MonoBehaviour:
   m_GameObject: {fileID: 851305182}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 986d20fa8e28249e3be84562e6061f00, type: 3}
+  m_Script: {fileID: 11500000, guid: a79441f348de89743a2939f4d699eac1, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  m_RenderShadows: 1
+  m_RequiresDepthTextureOption: 2
+  m_RequiresOpaqueTextureOption: 2
+  m_CameraType: 0
+  m_Cameras: []
+  m_RendererIndex: -1
+  m_VolumeLayerMask:
+    serializedVersion: 2
+    m_Bits: 1
+  m_VolumeTrigger: {fileID: 0}
+  m_VolumeFrameworkUpdateModeOption: 2
+  m_RenderPostProcessing: 0
+  m_Antialiasing: 0
+  m_AntialiasingQuality: 2
+  m_StopNaN: 0
+  m_Dithering: 0
+  m_ClearDepth: 1
+  m_AllowXRRendering: 1
+  m_AllowHDROutput: 1
+  m_UseScreenCoordOverride: 0
+  m_ScreenSizeOverride: {x: 0, y: 0, z: 0, w: 0}
+  m_ScreenCoordScaleBias: {x: 0, y: 0, z: 0, w: 0}
+  m_RequiresDepthTexture: 0
+  m_RequiresColorTexture: 0
+  m_Version: 2
+  m_TaaSettings:
+    quality: 3
+    frameInfluence: 0.1
+    jitterScale: 1
+    mipBias: 0
+    varianceClampScale: 0.9
+    contrastAdaptiveSharpening: 0
 --- !u!1 &1029994830
 GameObject:
   m_ObjectHideFlags: 0
@@ -1286,8 +1331,8 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 851305186}
-        m_TargetAssemblyTypeName: LevelsMenu, Assembly-CSharp
+      - m_Target: {fileID: 143991627}
+        m_TargetAssemblyTypeName: UINavigation, Assembly-CSharp
         m_MethodName: ToLevel
         m_Mode: 3
         m_Arguments:
@@ -1553,8 +1598,8 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 851305186}
-        m_TargetAssemblyTypeName: LevelsMenu, Assembly-CSharp
+      - m_Target: {fileID: 143991627}
+        m_TargetAssemblyTypeName: UINavigation, Assembly-CSharp
         m_MethodName: ToLevel
         m_Mode: 3
         m_Arguments:
