+++ Assets/Scenes/Levels/Level 1.unity
@@ -123,6 +123,103 @@ NavMeshSettings:
     debug:
       m_Flags: 0
   m_NavMeshData: {fileID: 0}
+--- !u!1001 &182517602
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    serializedVersion: 3
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 4482658631219995455, guid: d19c65b27b6a94f6896b85a04e025644, type: 3}
+      propertyPath: m_Name
+      value: PauseCanvas
+      objectReference: {fileID: 0}
+    - target: {fileID: 7327626698471134269, guid: d19c65b27b6a94f6896b85a04e025644, type: 3}
+      propertyPath: m_Pivot.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7327626698471134269, guid: d19c65b27b6a94f6896b85a04e025644, type: 3}
+      propertyPath: m_Pivot.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7327626698471134269, guid: d19c65b27b6a94f6896b85a04e025644, type: 3}
+      propertyPath: m_AnchorMax.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7327626698471134269, guid: d19c65b27b6a94f6896b85a04e025644, type: 3}
+      propertyPath: m_AnchorMax.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7327626698471134269, guid: d19c65b27b6a94f6896b85a04e025644, type: 3}
+      propertyPath: m_AnchorMin.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7327626698471134269, guid: d19c65b27b6a94f6896b85a04e025644, type: 3}
+      propertyPath: m_AnchorMin.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7327626698471134269, guid: d19c65b27b6a94f6896b85a04e025644, type: 3}
+      propertyPath: m_SizeDelta.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7327626698471134269, guid: d19c65b27b6a94f6896b85a04e025644, type: 3}
+      propertyPath: m_SizeDelta.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7327626698471134269, guid: d19c65b27b6a94f6896b85a04e025644, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7327626698471134269, guid: d19c65b27b6a94f6896b85a04e025644, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7327626698471134269, guid: d19c65b27b6a94f6896b85a04e025644, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7327626698471134269, guid: d19c65b27b6a94f6896b85a04e025644, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 7327626698471134269, guid: d19c65b27b6a94f6896b85a04e025644, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7327626698471134269, guid: d19c65b27b6a94f6896b85a04e025644, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7327626698471134269, guid: d19c65b27b6a94f6896b85a04e025644, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7327626698471134269, guid: d19c65b27b6a94f6896b85a04e025644, type: 3}
+      propertyPath: m_AnchoredPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7327626698471134269, guid: d19c65b27b6a94f6896b85a04e025644, type: 3}
+      propertyPath: m_AnchoredPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7327626698471134269, guid: d19c65b27b6a94f6896b85a04e025644, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7327626698471134269, guid: d19c65b27b6a94f6896b85a04e025644, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7327626698471134269, guid: d19c65b27b6a94f6896b85a04e025644, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+    m_RemovedGameObjects: []
+    m_AddedGameObjects: []
+    m_AddedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: d19c65b27b6a94f6896b85a04e025644, type: 3}
 --- !u!1 &321973117
 GameObject:
   m_ObjectHideFlags: 0
@@ -245,9 +342,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 25829bef4c00ba14fa9d98e502000cda, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  _color: {r: 0.509434, g: 0.12207189, b: 0.12207189, a: 0}
-  _size: 1
-  _isActive: 1
 --- !u!212 &1778920991
 SpriteRenderer:
   m_ObjectHideFlags: 0
@@ -321,3 +415,4 @@ SceneRoots:
   m_Roots:
   - {fileID: 321973120}
   - {fileID: 1778920992}
+  - {fileID: 182517602}
