+++ Assets/Scripts/UINavigation.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d57d1f6f24da81447925ea2f2cc37d0c
+guid: 14297e619d8364bdf8d2780dbd1c0762
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
